Move naka logic to raw tx and signed raw tx
