+++ crypto/src/main/java/org/web3j/crypto/RawTransaction.java
@@ -17,9 +17,13 @@ public class RawTransaction {
     private String to;
     private BigInteger value;
     private String data;
+    private String token;
+    private String exchanger;
+    private BigInteger exchangeRate;
 
-    protected RawTransaction(BigInteger nonce, BigInteger gasPrice, BigInteger gasLimit, String to,
-                           BigInteger value, String data) {
+    protected RawTransaction(BigInteger nonce, BigInteger gasPrice, 
+            BigInteger gasLimit, String to, BigInteger value, String data, 
+            String token, String exchanger, BigInteger exchangeRate) {
         this.nonce = nonce;
         this.gasPrice = gasPrice;
         this.gasLimit = gasLimit;
@@ -29,33 +33,41 @@ public class RawTransaction {
         if (data != null) {
             this.data = Numeric.cleanHexPrefix(data);
         }
+
+        this.token = token;
+        this.exchanger = exchanger;
+        this.exchangeRate = exchangeRate;
     }
 
-    public static RawTransaction createContractTransaction(
-            BigInteger nonce, BigInteger gasPrice, BigInteger gasLimit, BigInteger value,
-            String init) {
-
-        return new RawTransaction(nonce, gasPrice, gasLimit, "", value, init);
+    public static RawTransaction createContractTransaction(BigInteger nonce, 
+            BigInteger gasPrice, BigInteger gasLimit, BigInteger value,
+            String init, String token, String exchanger, 
+            BigInteger exchangeRate) {
+        return new RawTransaction(nonce, gasPrice, gasLimit, "", value, init, 
+            token, exchanger, exchangeRate);
     }
 
-    public static RawTransaction createEtherTransaction(
-            BigInteger nonce, BigInteger gasPrice, BigInteger gasLimit, String to,
-            BigInteger value) {
-
-        return new RawTransaction(nonce, gasPrice, gasLimit, to, value, "");
-
+    public static RawTransaction createEtherTransaction(BigInteger nonce, 
+            BigInteger gasPrice, BigInteger gasLimit, String to, 
+            BigInteger value, String token, String exchanger, 
+            BigInteger exchangeRate) {
+        return new RawTransaction(nonce, gasPrice, gasLimit, to, value, "", 
+            token, exchanger, exchangeRate);
     }
 
-    public static RawTransaction createTransaction(
-            BigInteger nonce, BigInteger gasPrice, BigInteger gasLimit, String to, String data) {
-        return createTransaction(nonce, gasPrice, gasLimit, to, BigInteger.ZERO, data);
+    public static RawTransaction createTransaction(BigInteger nonce, 
+            BigInteger gasPrice, BigInteger gasLimit, String to, String data, 
+            String token, String exchanger, BigInteger exchangeRate) {
+        return createTransaction(nonce, gasPrice, gasLimit, to, BigInteger.ZERO, 
+            data, token, exchanger, exchangeRate);
     }
 
-    public static RawTransaction createTransaction(
-            BigInteger nonce, BigInteger gasPrice, BigInteger gasLimit, String to,
-            BigInteger value, String data) {
-
-        return new RawTransaction(nonce, gasPrice, gasLimit, to, value, data);
+    public static RawTransaction createTransaction(BigInteger nonce, 
+            BigInteger gasPrice, BigInteger gasLimit, String to, 
+            BigInteger value, String data, String token, String exchanger, 
+            BigInteger exchangeRate) {
+        return new RawTransaction(nonce, gasPrice, gasLimit, to, value, data, 
+            token, exchanger, exchangeRate);
     }
 
     public BigInteger getNonce() {
@@ -80,5 +92,17 @@ public class RawTransaction {
 
     public String getData() {
         return data;
+    }
+
+    public String getToken() {
+        return token;
+    }
+
+    public String getExchanger() {
+        return exchanger;
+    }
+
+    public BigInteger getExchangeRate() {
+        return exchangeRate;
     }
 }
